+++ src/Keycloak.Net/Models/Clients/ClientPolicy.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace Keycloak.Net.Models.Clients
+{
+    public class ClientPolicy : Policy
+    {
+        [JsonProperty("clients")]
+        public IEnumerable<string> Clients { get; set; }
+    }
+}
