+++ src/Keycloak.Net/Models/Clients/RolePolicy.cs
@@ -5,52 +5,16 @@ using Newtonsoft.Json;
 
 namespace Keycloak.Net.Models.Clients
 {
-    public class Policy
+    public class RolePolicy : Policy
     {
-        [JsonProperty("id")]
-        public string Id { get; set; }
-
-        [JsonProperty("name")]
-        public string Name { get; set; }
-
-        [JsonProperty("description")]
-        public string Description { get; set; }
-
         [JsonConverter(typeof(PolicyTypeConverter))]
-        public PolicyType Type { get; set; }
-
-        [JsonConverter(typeof(PolicyDecisionLogicConverter))]
-        public PolicyDecisionLogic Logic { get; set; } 
-
-        [JsonConverter(typeof(DecisionStrategiesConverter))]
-        public DecisionStrategy DecisionStrategy { get; set; }
-
-        [JsonProperty("config")]
-        public PolicyConfig Config { get; set; }
-    }
-
-    public class RolePolicy
-    {
-        [JsonProperty("id")]
-        public string Id { get; set; }
-
-        [JsonProperty("name")]
-        public string Name { get; set; }
-
-        [JsonProperty("description")]
-        public string Description { get; set; }
-
-        [JsonConverter(typeof(PolicyTypeConverter))]
-        public PolicyType Type { get; set; } = PolicyType.Role;
-
-        [JsonConverter(typeof(PolicyDecisionLogicConverter))]
-        public PolicyDecisionLogic Logic { get; set; } 
-
-        [JsonConverter(typeof(DecisionStrategiesConverter))]
-        public DecisionStrategy DecisionStrategy { get; set; }
+        public new PolicyType Type { get; set; } = PolicyType.Role;
 
         [JsonProperty("roles")]
         public IEnumerable<RoleConfig> RoleConfigs { get; set; }
+        
+        [JsonProperty("config")]
+        public PolicyConfig Config { get; set; }
     }
 
     public class RoleConfig
@@ -60,17 +24,6 @@ namespace Keycloak.Net.Models.Clients
 
         [JsonProperty("required")]
         public bool Required { get; set; }
-    }
-
-    public enum PolicyType
-    {
-        Role,
-        Client,
-        Time,
-        User,
-        Aggregate,
-        Group,
-        Js
     }
 
     public class PolicyConfig
